+++ src/dist.h
@@ -16,7 +16,7 @@ typedef struct setResult
 	double jaccard;
 } setResult_t;
 
-setResult_t getJaccard(vector<uint64_t> list1, vector<uint64_t> list2);
+setResult_t getJaccard(vector<uint32_t> list1, vector<uint32_t> list2);
 
 
 void tri_dist(vector<sketch_t> sketches, string outputFile, int kmer_size, double maxDist, int numThreads);
